commit tasks status as a separate model file
